requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
google-cloud-storage>=2.0.0
fsspec>=2023.1.0
gcsfs>=2023.1.0
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0